Add ability to hide exam after due date
TNL-4366

This commit adds a new field, `hide_after_due` to the ProctoredExam model. This field
is intended to allow course authors to override the default setting and keep exam
results hidden from learners after the due date for the exam has passed.

Also included are migrations, tests, and api updates to allow this functionality to
be used.