Record architecture decision records
this process change is motivated by the need to document and openly
discuss the architectural decisions made in the course of refactoring
edx-proctoring to allow pluggable proctoring providers. This will
encourage us to document these decisions in a way that encodes context
of the context in which these decisions were made.