Mock _get_payload properly
The original code used a singleton, so you could get the backend, patch
it, and then get it again, and be running the patched object.  Now
get_backend_provider returns a new object each time, so the product code
uses a new unpatched object, so the assert code wasn't being called.

Also added assertions that the mocks were being called.