Improve screen real estate usage of proctoring courseware banner
The banner took up too much room with the browser
zoomed in. This adds a button to hide/show much of the copy, reducing the
portion of the page on zoomed-in desktop browsers stickily obscuring
courseware. The toggle's initial state is set according to browser
width.

Also updates some sematic structural details

JIRA:EDUCATOR-3899
JIRA:EDUCATOR-3900