Prevent negative exam durations
In corner cases in which users hold onto an intermediate proctoring
session state over the deadline for starting an exam, it's possible
the time still remaining to them may flip into the red. In that case,
we want to normalize it back to 0, so the time displayed to course
teams doesn't seem physically impossible.

JIRA:EDUCATOR-3993