Update serialization error test.
The test has had to change a few times due to the fact that the
underlying implementation of DRF continues to change exactly what the
error response looks like.  Update it so that we ensure that the
failures we expect are there but stop caring about the exact content of
the underlying message.