Update the version numbers to prepare for a release.
Making this a major version update since we will be un-pinning DRF on
the python side.  Unpinning it in our setup.py may result in it getting
upgraded for upstream dependencies that didn't explicitly pin it.

This might break the upstream project so we want to call that out with a
major version update.