Don't run JS tests on every run.
The way the travis file was written, travis would run tox(which would run
the python tests) and then it would run the js tests.

So the JS tests would be run 10 times.  Instead we give it its own test
context so that it only runs once and tests finish much faster.