MST-734 Add due_date to the learning_sequence accessible_sequences
Before this change, the date we use to test the sequence visibility was
max datetime. This means the onboarding exams with a due date will not
allow the onboarding panel to show. With this commit, the onboarding
status panel will show to learners before or after the due date of the
exam.