fix: Fix JWT Encoding Bug with Version 2.0.1 of pyjwt[crypto] Library
Version 2.0.1 of the pyjwt library includes a change to the jwt.encode(...) return type. The method now returns tokens as strings instead of byte strings. In the backend base class BaseRestProctoringProvider, code was originally decoding the token as utf-8. When using version 2.0.1 of the pyjwt[crypto] library, this code was raising a "AttributeError: 'str' object has no attribute 'decode'" error. This code fixes this behavior to treat the return type of jwt.encode(...) as a string.