fix: process onboarding exams correctly in onboarding status API used by onboarding panel
When returning information about a learner's onboarding status in a course, we exclude onboarding exams that the Learning Sequences API marks as inaccessible. While processing the list of onboarding exams in the course, the list was being modified during iteration, leading to incorrect behavior. This code change fixes this behavior. This fix also exposed a bug in the corresponding tests, which was also fixed in this code change.