fix: Fix bug in StudentOnboardingStatusView for no sequence schedule for a sequence
This code change fixes a bug introduced by https://github.com/edx/edx-proctoring/pull/855 ([MST-746](https://openedx.atlassian.net/browse/MST-746)). The Learning Sequences API returns details about a given user's course outline, including sequence scheduling information. In some cases, a sequence that is inaccessible to a learner is still returned in the schedule, because the learner is allowed to know of the sequence's existence. A future onboarding exam is an example of this. In other cases, inaccessible sequences should be truly inaccessible to the learner, and, as such, they are not returned in the schedule. An onboarding exam that is content gated by enrollment track to an enrollment track in which the learner is not enrolled is an example of this. The bug was that the StudentOnbaordingStatusView was not taking into account that an inaccessible sequence might not have a schedule. This code change fixes that bug by adding None checking.