fix: exam content will not be viewable if due date has passed
Previously, if a learner had submitted their proctored exam and if the
exam due date had passed, they were able to acknowledge their status
and view the exam content. This is no longer wanted and poses an integrity
threat. Students will not be able to acknowledge their status after this change,
and exam content will only be viewable if the django setting PROCTORED_EXAM_VIEWABLE_PAST_DUE
is set to True.