fix: Fix JSON Parse Error for Error Responses without JSON Parseable Text
In pull #879 (https://github.com/edx/edx-proctoring/pull/879), error handling was added to the Javascript that renders the Student Onboarding Status panel in the Instructor Dashboard. This error handling parsed the JSON returned by the response from the StudentOnboardingStatusByCourseView endpoint. However, this fails if the data returned in the response is not valid JSON. The changes in this commit address this bug by adding error handling to the call to parse the JSON response. If the JSON is not valid JSON, then a generic error message is returned and displayed by the frontend.