fix: Fix bug with user IDs returned by proctoring vendor for learners not enrolled in eligible mode
This pull request fixes a bug that occurs when a proctoring vendor returns onboarding information that includes user IDs that represent learners that are not returned by the edX API as being enrolled in the course in a proctoring eligible mode. It also adds a logging statement to enable further investigation.