fix: restore verified name waffle check to proctoring
This flag was read and passed from the front end but we've removed
that. If verified names are in use we still want to pass them to the
proctoring and use them in signals when available. Ask name
verification, if it is in use, whether it is enabled rather than
directly engaging with its feature flag.

MST-1020