feat: management command for triggering a status update to a proctored exam attempt through the attempt review
As a result of https://github.com/edx/edx-platform/pull/28559, the signal and handler connection in edx-proctoring was broken. As a result, all incoming reviews for proctored exam attempts were unable to correctly update the status of the attempt. This management command is needed so that we can rerun the post_save signal handler for proctored exam attempt reviews, so that the status for those attempts can be updated properly.