feat: use video review url provided by PSI for instructor dashboard url
The instructor dashboard url is passed to zendesk tickets sent to support when a review for an attempt is received. This link has never worked for PSI attempts, as PSI's backend class is not set up to support the instructor dashboard view. With this PR, PSI's backend class is built out to work with the instructor dashboard view (which requires a backend class to provide a link to redirect to). PSI's backend class provides the attempt video review url as the link to redirect to.