fix: allow practice exam attempts to always trigger grade and credit updates
Practice exams should always be allowed to trigger grade and credit updates, as a singular approved practice attempt means that a user has passed, as opposed to the
cumulative approach that is taken for non practice proctored exams. In order to fix the credits and grades for a particular exam, a management command has been added
to re-trigger grade and credit updates.