fix: timed exams should remain visible after course end date
In a self-paced course, if a learner has submitted a timed exam and the self-paced due date has passed,
the exam content becomes visible. Once the course end date has passed, however, that content is hidden again.
This creates an inconsistent experience for the learner, and support has asked that we allow exam content to
remain visible after the course end date.