fix: exam attempt should be resumable if it has been marked as ready to resume
MST-1124. If an exam attempt has been previously marked as ready to resume, it should
remain resumable by a learner regardless of whether or not they receive a review. In order
to do this, resume states should no longer be represented as statuses, but instead boolean
fields on the attempt model.