fix: do not use course's due date when a timed exam has one specified
24199b466 introduced course end date check to allow viewing timed exams after
course end date in self-paced courses. However, this change also means that
the students cannot submit exams once the course due date is in the past.
The exam's due date should have a higher priority, so that the students with
an individual due date extension (set through the Instructor Dashboard) can
still complete an exam in such a case.