fix: remove logilab-common requirement
The tox quality environment is used to run quality tests (i.e. pylint, pycodestyle, and isort) on the codebase. This environment installs dependencies from both the test and quality requirement files.

This has caused a dependency resolution error with the importlib-metadata dependency. In the test requirements, importlib-metadata is a dependency of the logilab-common dependency, and this installs importlib-metadata=6.11.0. In the quality requirements, importlib-metadata is a dependency of the rstcheck and rstcheck-core dependencies, and this installs importlib-metadata=7.0.1. This causes a dependency resolution conflict when running the tox quality environment, because both the test and quality requirements files are installed in this environment, and versions 6.11.0 and 7.0.1 of importlib-metadata are requested, respectively.

The simplest solution is to remove the logilab-common dependency from the test requirements file, as it has no obvious purpose, particularly since this project does not explicitly install importlib-metadata and logilab-common, rstcheck, and rstcheck-core have very loose requirements ranges. This means we cannot easily ensure that the two importlib-metadata package versions match by modifying our versions of logilab-common, rstcheck, or rstcheck-core. It was also not possible to remove the test requirements from the tox quality environment, because then the quality checks fail with import errors for modules used in the tests (e.g. mock).