fix: clear all ProctoredExamSoftwareSecureReviewHistory instances of PII
This commit changes the time at which we read instances of the ProctoredExamSoftwareSecureReviewHistory model. This is necessary because the act of saving the ProctoredExamSoftwareSecureReview model causes the pre-save instance to be saved to the ProctoredExamSoftwareSecureReviewHistory table. Because we read instances of the ProctoredExamSoftwareSecureReviewHistory model prior to calling save on the instances of the ProctoredExamSoftwareSecureReview model, we do not have the full queryset we need. Reading the instances of the ProctoredExamSoftwareSecureReviewHistory model from the database after we clear all the instances of the ProctoredExamSoftwareSecureReview model fixes this issue.