perf: add pre_cache_exams_for_course API call
This is to address the performance issue when outline page loading
in Studio causes multiple calls to get_exam_by_content_id() for
every subsection in the course. On one test course, this resulted
in about a hundred separate queries, despite the fact that only two
of the subsections had corresponding ProctoredExam models.